backend/requirements.txt
fastapi
uvicorn
sqlmodel
mysql-connector-python
python-dotenv
